Add files via upload
Tracks time spent on each KDE activity (KDE Activities). 
Time tracking for each activity works automatically, without the need for any action when switching between activities.

This is a working version, but there are still things to be done.

I'm no longer working on it.

Feel free to copy and improve the project.